feat: Add finary creation functionality
Remove dependency on hard coded LOOK_UP dictionary and replace with self-populating
dictionary functionality.

If a £ is present in the text file to be converted, it seems to add an Â before that
character and I am unsure why. I will try and investigate this further, but I am happy
that the main functionality is working and the bug should not impact on many files...
Unless you start to use it for your accounts!

I have not tested all characters so there may be other erroneous characters being added.